Don't restore previous color value if it was explicitly deselected
If you deselect a color value before setting the width to zero, when
the width is changed to a non-zero value that color should not be
restored (because it was explicitly deselected). This was happening
because the state that tracked previous values would not update when
the value was set to undefined.